Clarify benign CNAME usage
Several news articles have been pointing to this documentation as proof that Adobe is using CNAMEs to circumvent consumer privacy protections, because some of the text was carelessly worded.  This text has been updated to make it clear that Adobe respects consumer privacy and only supports customer use of CNAMEs for their own tracking of their customers.